New facility added to specify in a config file that simulated damaged content be incorrectly written to a cache, rather than generated as damaged.
Content is damaged by setting every 10th byte to 0.  The first time the content
is written to the repository, it is written in damaged form.  After that it is
written normally.

To specify a node to be marked as damaged, use the "badCachedFileLoc" and
"badCachedFileNum" config file settings, in the same format as the existing
"badFileLoc" and "badFileNum" settings, which continue to operate as before.

No substantive changes were made to GoslingCrawlerImpl; just whitespace cleanups
to make it more readable.