The main change for this routine:
DiskVoteBlocks had some code that was supposed to speed up retrieval when retrieval was in order.  In practice, the speed-up could give the wrong value.

I've added a test specific to the error in the speed-up, as well as a few extra tests (just because.)

IMPORTANT NOTE: I have changed V3TestUtils so that "makeVoteBlockList" returns a vote block list with different-sized entries.  This change is important; not having it will obscure many tests.