Basic building blocks for supporting crawling journals from www.thieme-connect.de/ejournals.
Here is a brief overview of the design:
1. All the journals are hosted on the thieme connect platform which is understood by the ThiemeConnectPlugin. Each journal has a unique string identifier (ajp for 'American Journal of Perinatology for example). The example journals crawled by this change list can be found in thieme_connect_ejournals.tdb file.
2. Each year in a given journal represents an Archival unit (AU). Due to AJAXy architecture of the website, there is no HTML toc page containing all the issues in a given AU. Instead, on selecting a given year, an AJAX request of following form is sent to server:
https://www.thieme-connect.de/ejournals/json/issues?shortname=ajp&year=2008
This url results in a json response from server. The link extraction from this response is handled by ThiemeJsonLinkExtractor.
3. The actual start_url of this plugin is not the url above because the server requires a valid session cookie to respond to this request. Instead, we use a common url for each journal like following:
https://www.thieme-connect.de/ejournals/toc/ajp
This url results in an HTML page containing the list of latest issues for this journal. By definition of an AU, we can not include links extracted from this page if they belong to a different AU. We therefore have a special ThiemeHtmlLinkExtractor to embed this business logic. If this extractor encounters the start url page, it will only emit the json url above otherwise, it will delegate to Gosling extractor.
4. For testing purposes only, we change the lockss permission to allow any web page to pass the check.