Implemented feature 3741: Option to display health rating in List*Holdings.
New and updated tests; copyright notices, formatting.

Parameterized the ListUtil.fromArray() method.

KbartTitleHealthWrapper is a new wrapper class which adds a health rating to a KBART title record.

KbartConverter.createKbartTitlesWithRanges() maps created KbartTitles to the list of TdbAus which underlie them. This is now the main
createKbartTitles method.

AuHealthMetric has new methods to get health, to aggregate health over a range of AUs, and to calculate the threshold for including an
AU as "preserved".

Added max health value to PreservationStatus, and expected class to HealthMetrics. Sourced more metrics and added
interpretation.